rename ContactSection component (was Contact) to differetiate from the new ContactForm component
